fix: address local code review findings
- Convert ChimeDedup/OnceDedup from __init__ to __new__ (PY-CC-1)
- Add __slots__ to ChimeResolver (stateless class)
- Add missing re-exports to __init__.py (_install_token_redact_filter,
  _DEDUP_WINDOW_SECONDS)
- Move config tests to tests/test_voxd_config.py (PL-PL-3)

init_violations on dedup.py: 2→0
