+++ .oo-baseline.json
@@ -434,7 +434,7 @@
     "avg_params": 0.0,
     "max_complexity": 0.0,
     "avg_complexity": 0.0,
-    "module_size": 106.0,
+    "module_size": 110.0,
     "classes_per_module": 0.0,
     "class_to_func_ratio": 1.0,
     "init_violations": 0.0,
@@ -460,7 +460,7 @@
     "avg_params": 0.5,
     "max_complexity": 4.0,
     "avg_complexity": 2.5,
-    "module_size": 42.0,
+    "module_size": 43.0,
     "classes_per_module": 1.0,
     "class_to_func_ratio": 1.0,
     "init_violations": 0.0,
@@ -486,10 +486,10 @@
     "avg_params": 1.0,
     "max_complexity": 8.0,
     "avg_complexity": 3.2,
-    "module_size": 165.0,
+    "module_size": 175.0,
     "classes_per_module": 3.0,
     "class_to_func_ratio": 1.0,
-    "init_violations": 2.0,
+    "init_violations": 0.0,
     "public_attr_violations": 0.0,
     "future_annotations": 1.0
   },
